projects/project3_1/requirements.txt
# FAQ检索系统核心依赖包
# 核心框架
llama-index-core==0.10.68.post1
llama-index-vector-stores-faiss==0.1.2
llama-index-llms-dashscope
llama-index-embeddings-dashscope
dashscope

# 向量数据库
faiss-cpu>=1.9.0

# 对话管理
langgraph

# 数据处理
pandas
numpy

# 日志和配置
python-dotenv
loguru

# HTTP客户端
httpx

# Web API框架
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
